Rename `rustc_codegen_llvm/errors.rs` into `rustc_codegen_llvm/diagnostics.rs`
